% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-country.R
\docType{data}
\name{country}
\alias{country}
\title{Country Data}
\format{
A data frame with 167 observations and 10 variables:
\describe{
\item{country}{country name}
\item{child_mort}{Death of children under 5 years of age per 1000 live births}
\item{exports}{Exports of goods and services per capita. Given as percentage of the GDP per capita}
\item{health}{Total health spending per capita. Given as percentage of GDP per capita}
\item{imports}{Imports of goods and services per capita. Given as percentage of the GDP per capita}
\item{income}{Net income per person}
\item{inflation}{The measurement of the annual growth rate of the Total GDP}
\item{life_expec}{The average number of years a new born child would live if the current mortality patterns are to remain the same}
\item{total_fer}{The number of children that would be born to each woman if the current age-fertility rates remain the same.}
\item{gdpp}{The GDP per capita. Calculated as the Total GDP divided by the total population.}
}
}
\source{
Kaggle Datasets, https://www.kaggle.com/datasets/rohan0301/unsupervised-learning-on-country-data
}
\usage{
country
}
\description{
Country data with socio-economic and health factors that determine the overall development of the country.
}
\examples{
country
name_country <- country$country

}
\keyword{datasets}
